abstract droppability to Droppable
